% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{mod}
\alias{mod}
\title{Convert API datetime strings to <dttm> format}
\usage{
mod(x)
}
\arguments{
\item{x}{datetime character string}
}
\value{
dttm object
}
\description{
Works with AWS and Github at this stage
}
